refactor: Add base Application class and inherit in other source sets
