+++ app/src/release/java/me/worric/souvenarius/SouvenirApp.java
@@ -1,45 +1,8 @@
 package me.worric.souvenarius;
 
-import android.app.Activity;
-import android.app.Application;
-import android.app.Service;
-
-import com.jakewharton.threetenabp.AndroidThreeTen;
-
-import javax.inject.Inject;
-
-import dagger.android.AndroidInjector;
-import dagger.android.DispatchingAndroidInjector;
-import dagger.android.HasActivityInjector;
-import dagger.android.HasServiceInjector;
-import timber.log.Timber;
-
-public class SouvenirApp extends Application implements HasActivityInjector, HasServiceInjector {
-
-    @Inject
-    DispatchingAndroidInjector<Activity> mActivityDispatcher;
-    @Inject
-    DispatchingAndroidInjector<Service> mServiceDispatcher;
-
-    @Override
-    public void onCreate() {
-        super.onCreate();
-        Timber.plant(new Timber.DebugTree());
-        AndroidThreeTen.init(this);
-        DaggerSouvenirAppComponent.builder()
-                .application(this)
-                .build()
-                .inject(this);
-    }
-
-    @Override
-    public AndroidInjector<Activity> activityInjector() {
-        return mActivityDispatcher;
-    }
-
-    @Override
-    public AndroidInjector<Service> serviceInjector() {
-        return mServiceDispatcher;
-    }
-
+/**
+ * Specialization of the custom Application base class. It is intentionally left empty, as there are
+ * no customizations for the release build type, but nevertheless required for the Manifest entry
+ */
+public class SouvenirApp extends BaseSouvenirApp {
 }
